making auth requests to the app

+++ frontend/src/components/WorkoutDetails.js
@@ -1,16 +1,25 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from "../hooks/useAuthContext";
 
-function WorkoutDetails({workout}){
-    const {dispatch }  = useWorkoutContext()
+function WorkoutDetails({ workout }) {
+    const { dispatch } = useWorkoutContext()
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
-        const response = await fetch('/api/workouts/'+ workout._id, {
-            method: 'DELETE'
+        if (!user){
+            return
+        }
+
+        const response = await fetch('/api/workouts/' + workout._id, {
+            method: 'DELETE',
+            headers: {
+                'authorization': `bearer ${user.token}`
+            }
         })
         const json = await response.json()
-        if (response.ok){
-            dispatch({type: 'DELETE_WORKOUT', payload: json})
+        if (response.ok) {
+            dispatch({ type: 'DELETE_WORKOUT', payload: json })
         }
     }
     return (
@@ -18,7 +27,7 @@ function WorkoutDetails({workout}){
             <h4>{workout.title}</h4>
             <p><strong>Load (kg): </strong>{workout.load}</p>
             <p><strong>Reps: </strong>{workout.reps}</p>
-            <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
+            <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
             <span className="material-symbols-outlined" onClick={handleClick}>Delete</span>
         </div>
     )
